+++ src/App.js
@@ -2,9 +2,9 @@ import React, { useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import Cookies from 'js-cookie';
-import Register from './views/Register';
+import Register from './views/user/Register';
 import Login from './views/user/Login';
-import Verify from './views/Verify';
+import Verify from './views/user/Verify';
 import { setToken, setAuthenticated } from './redux/authslice';
 
 function App() {
@@ -20,7 +20,7 @@ function App() {
 
   return (
     <Routes>
-      <Route path="/" element={<Register />} />
+      <Route path="/register" element={<Register />} />
       <Route path="/login" element={<Login />} />
       <Route path="/verify" element={<Verify />} />
     </Routes>
