+++ src/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import authReducer from './slices/auth';
-import messageReducer from './slices/message';
+import authReducer from './redux/authslice';
+import messageReducer from './redux/messageSlice';
 
 const reducer = { auth: authReducer, message: messageReducer };
