+++ src/views/user/Login.js
@@ -6,6 +6,7 @@ import { GrFacebook } from 'react-icons/gr';
 import { FcGoogle } from 'react-icons/fc';
 import * as Yup from 'yup';
 import { useDispatch } from 'react-redux';
+import { Link } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import SocialButton from '../../components/SocialButton';
 import { login } from '../../utils/api';
@@ -40,12 +41,15 @@ export default function Login() {
         <div className="w-8/12 md:w-6/12 bg-white p-16 ">
           <div>
             <p className=" py-4 font-bold text-xl md:text-3xl">Sign In</p>
+            <Link to="/register">
+
             <p>
-              Dont have an account?
+              Don't have an account?
               <span className="text-orange-dark font-semibold ml-1">
                 Sign Up
               </span>
             </p>
+            </Link>
           </div>
           <div>
             <Formik
