+++ src/views/user/Register.js
@@ -7,7 +7,7 @@ import {
   Formik, Field, Form, ErrorMessage,
 } from 'formik';
 import * as Yup from 'yup';
-import { register } from '../slices/auth';
+import { register } from '../../redux/authslice';
 
 export default function Register() {
   const navigate = useNavigate();
